# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "Pydantic model for an RFC 7807 problem-details payload.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Human-readable explanation of this specific occurrence, distinct from the generic ``title``.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "URI of the specific occurrence (typically the request path); helps correlate with logs.",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code repeated in the body so the payload is self-contained for clients that drop the status line.",
            "maximum": 599,
            "minimum": 100,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, stable summary of the problem.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI reference identifying the problem class. Relative URIs resolve under the API host; use ``about:blank`` for generic.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "contact@protea.example.org",
      "name": "PROTEA Team"
    },
    "description": "**PROTEA** \u2014 Protein Representation and Ontology-Term Enrichment Analysis.\n\nManages the full pipeline from UniProt sequence ingestion through GPU embedding computation (ESM-2, ESM3c, T5) to KNN-based GO term prediction.\n\nAll long-running operations are queued via RabbitMQ and tracked as `Job` rows with a full event audit trail. Use `GET /jobs/{id}/events` to stream real-time progress.",
    "title": "PROTEA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "description": "List jobs with optional filtering.\n\nBy default only top-level jobs (no parent) are returned. Set\n``include_children=true`` or filter by ``parent_job_id`` to see batch\nsub-jobs from distributed pipelines. Filters travel as discrete query\nparameters on the wire; the dependency bundles them into\n:class:`JobListFilters` for the handler.",
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "description": "Filter by ``JobStatus`` enum value (``queued``, ``running``, ``succeeded``, ``failed``, ``cancelled``). Unknown values return 400.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by ``JobStatus`` enum value (``queued``, ``running``, ``succeeded``, ``failed``, ``cancelled``). Unknown values return 400.",
              "title": "Status"
            }
          },
          {
            "description": "Filter by registered operation name (e.g. ``compute_embeddings``, ``predict_go_terms``).",
            "in": "query",
            "name": "operation",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by registered operation name (e.g. ``compute_embeddings``, ``predict_go_terms``).",
              "title": "Operation"
            }
          },
          {
            "description": "When true, include batch sub-jobs (rows with a non-null ``parent_job_id``). Default false keeps the list focused on top-level work.",
            "in": "query",
            "name": "include_children",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, include batch sub-jobs (rows with a non-null ``parent_job_id``). Default false keeps the list focused on top-level work.",
              "title": "Include Children",
              "type": "boolean"
            }
          },
          {
            "description": "Filter to sub-jobs of a specific parent (overrides ``include_children``). Useful for inspecting one batch.",
            "in": "query",
            "name": "parent_job_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to sub-jobs of a specific parent (overrides ``include_children``). Useful for inspecting one batch.",
              "title": "Parent Job Id"
            }
          },
          {
            "description": "Max rows per page; capped at 500.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Max rows per page; capped at 500.",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Cursor for pagination; return rows with ``created_at < after`` only. Use the ``created_at`` of the last row from the previous page to walk forward.",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor for pagination; return rows with ``created_at < after`` only. Use the ``created_at`` of the last row from the previous page to walk forward.",
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Jobs V1 Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Create a Job row and publish its ID to the specified RabbitMQ queue.\n\nThe job transitions `QUEUED \u2192 RUNNING \u2192 SUCCEEDED/FAILED` as the worker processes it.\nUse `GET /jobs/{id}/events` to poll structured progress events in real time.",
        "operationId": "create_job_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Job V1 Jobs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create and enqueue a job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "delete": {
        "description": "Permanently delete a job and its event log. Running jobs cannot be deleted (409).",
        "operationId": "delete_job_v1_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Job V1 Jobs  Job Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "Retrieve full details for a single job including its payload, meta, and progress counters.",
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job V1 Jobs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get job details",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "description": "Mark a job (and any non-terminal child jobs) as CANCELLED.\n\nAlready-finished jobs (SUCCEEDED/FAILED) are returned as-is with no state change.\nChildren in QUEUED are cancelled immediately.  Children in RUNNING are also\nmarked CANCELLED; the worker's parent-check in BaseWorker.handle_job() will\ndetect the cancelled parent on the next iteration and stop gracefully.",
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Job V1 Jobs  Job Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel a job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
...
}